+++ Vector3.cpp
@@ -2,7 +2,7 @@
 
 Vector3::Vector3() : x(0), y(0), z(0) {};
 
-Vector3::Vector3(float x, float y, float z) {
+Vector3::Vector3(double x, double y, double z) {
     this->x = x;
     this->y = y;
     this->z = z;
@@ -40,25 +40,25 @@ Vector3& Vector3::operator-(const Vector3 &rhs) {
     return *this;
 }
 
-Vector3& Vector3::operator+(float scalar) {
+Vector3& Vector3::operator+(double scalar) {
     this->x += scalar ;
     this->y += scalar ;
     this->z += scalar ;
     return *this;
 }
-Vector3& Vector3::operator-(float scalar) {
+Vector3& Vector3::operator-(double scalar) {
     this->x -= scalar ;
     this->y -= scalar ;
     this->z -= scalar ;
     return *this;
 }
-Vector3& Vector3::operator*(float scalar) {
+Vector3& Vector3::operator*(double scalar) {
     this->x *= scalar ;
     this->y *= scalar ;
     this->z *= scalar ;
     return *this;
 }
-Vector3& Vector3::operator/(float scalar) {
+Vector3& Vector3::operator/(double scalar) {
     this->x /= scalar ;
     this->y /= scalar ;
     this->z /= scalar ;
